refactor smd
